requirements.txt
spyne
zeep
lxml
mysql-connector-python
PyJWT
python-dotenv
kafka-python